Move job mapper under services folder
