+++ src/JobsInFinland.Api.Productizer/Services/IJobPostingMapper.cs
@@ -1,7 +1,7 @@
 using JobsInFinland.Api.Infrastructure.CodeGen.Model;
 using JobsInFinland.Api.Productizer.Models.Testbed;
 
-namespace JobsInFinland.Api.Productizer;
+namespace JobsInFinland.Api.Productizer.Services;
 
 public interface IJobPostingMapper
 {
